V1.4.8 (#58)
* don't remove pwfile on dryrun
* list or -- or -
* fix missing --network option for rules
* fix pname use

+++ bin/tdh-masters-init.sh
@@ -438,7 +438,7 @@ for name in $names; do
     echo ""
 done
 
-if [ -e "$pwfile" ]; then
+if [ -e "$pwfile" ] && [ $dryrun -eq 0 ]; then
     ( rm $pwfile )
 fi
